third_party/WebKit/LayoutTests/external/wpt/web-animations/interfaces/KeyframeEffect/composite-expected.txt

This is a testharness.js-based test.
PASS Default value
PASS Change composite value
FAIL Unspecified keyframe composite value when setting effect composite assert_equals: unspecified keyframe composite value should be absent even if effect composite is set expected (undefined) undefined but got (string) "replace"
PASS Specified keyframe composite value when setting effect composite
Harness: the test ran to completion.
